refactor(BA-7820): declare the membership edge as an entity type
The edge names two virtual entity nodes and belongs to neither, and its cap rows
hang off it, so it is an entity of its own rather than a field of either end.
EntityMembershipID moves from a plain NewType in data/permission/id.py to an
EntityIdentifier answering EntityMembershipEntityType, which is what lets
GroupMembershipNotFoundError name the row it reads: it now reports
entity-membership_read_not-found instead of group_read_not-found. The virtual
entity node itself stays out of the catalog.

Nothing outward moves with it. AuthService.get_role is the error's only consumer
and rethrows it as ObjectNotFound, so no client has ever seen the old code.
