+++ src/ai/backend/common/data/permission/id.py
@@ -3,7 +3,6 @@ from uuid import UUID
 
 __all__ = (
     "EntityMembershipCapID",
-    "EntityMembershipID",
     "FieldPath",
 )
 
@@ -11,5 +10,4 @@ __all__ = (
 # A path covers its descendants.
 FieldPath = NewType("FieldPath", str)
 
-EntityMembershipID = NewType("EntityMembershipID", UUID)
 EntityMembershipCapID = NewType("EntityMembershipCapID", UUID)
